userLocation moved to services
